Use `jldopen` in `collect_results!` when scanning JLD2 files.
Employ lazy loading via Requires.jl to allow special load functions for particular file types in `collect_results!`, and falling back to `wload` as a default. This potentially makes scanning collections of large files much faster when using the `black_list` argument of `collect_results!`.

An example use case are simulations that generate amounts of data where the parameters are stored in a Dict, together with the data under the keyword `data`, all saved in a JLD2 file. If we only want to enumerate all the parameters that have been run, we would use `collect_results(datadir("simulations"), black_list = ("data",))`. With `wload`, this still loads the whole file into memory before ignoring the `data` key. With `jldopen`, only the parameters are loaded into memory. In my tests, this leads to around 400x speed up when scanning large collections of files.

For now, only `jldopen` for JLD2 files is implemented, but this could potentially be extended to other data storage formats that allow memory mapped access to storage keys.