Add `wload kwargs` to `produce_or_load` (#412)
* Add wload_kwargs to produce_or_load

An additional keyword argument called wload_kwargs was added to
produce_or_load. This makes it possible to pass arguments to wload. For
example, if you need to load some data type from disk but the type
definition has changed, you need to provide JLD2 a typemap to rebuild
the original object. This makes it possible to pass the typemap to wload
via produce_or_load. wload_kwargs defaults to an emtpy named tuple.

The default type wsave_kwargs was changed to an empty named tuple. This
seems like a more natural choice than as an empty Dict which would only
work with kwarg splatting if all the keys happened to be of type Symbol. Also
slurping kwargs gives a named tuple.

* update CHANGELOG and Project.toml version