implement custom load function for collect_results
This adds an optional wrapper around the loading of data when being
processed by collect_results. This is useful in case the data being
loaded from file needs to be modified before being processed by
collect_results. The wrapper can be passed to collect_results using the
new kwarg `load_function_wrapper`.

For example, consider you have a simulation result that is saved to disc
as a struct. When loaded from disc it will be as a one-element dict
containing that struct. The struct may contain many fields that should
form the columns of the DataFrame. In that case you'd want to convert
the struct to a dictionary after being loaded but before being processed
by collect_results. The `load_function_wrapper` provides such a
possibility.

By default `load_function_wrapper=nothing` and the default methods will
be used to load data (so `wload` and `jld2.jldopen`).