Avoid deadlock when reading stdout and stderr from commands
The `gitpatch` function sometimes hangs for large git diffs. This seems to be
happening because the `read_stdout_stderr` function is writing to the `out` and
`err` pipes, but no task is reading from them, so they fill up, and the process
locks up waiting to write more data.

This change spawns separate tasks to read from the `out` and `err` pipes, thus
avoiding the deadlock.

The Julia compat for this package is set to 1.0, so I opted to use `@async`
instead of `Threads.@spawn`.

Fixes #300